flatten src
