+++ package.json
@@ -27,6 +27,14 @@
     {
       "name": "Philippe Rivière",
       "url": "https://visionscarto.net"
+    },
+    {
+      "name": "Enrico Spinielli",
+      "url": "https://enrico.spinielli.net/"
+    },
+    {
+      "name": "Ronnie Bathoorn",
+      "url": "https://github.com/bathoorn"
     }
   ],
   "type": "module",
