+++ src/dodecahedral.js
@@ -1,5 +1,5 @@
-import {acos, asin, degrees, sqrt} from "../math.js";
-import voronoi from "./voronoi.js";
+import {acos, asin, degrees, sqrt} from "./math.js";
+import voronoi from "./polyhedral/voronoi.js";
 
 export default function() {
   const A0 = asin(1/sqrt(3)) * degrees;
