+++ src/index.js
@@ -1,11 +1,11 @@
 export {default as geoClipPolygon} from "./clip/polygon.js";
 export {default as geoIntersectArc} from "./intersect.js";
 export {default as geoPolyhedral} from "./polyhedral/index.js";
-export {default as geoPolyhedralButterfly} from "./polyhedral/butterfly.js";
-export {default as geoPolyhedralCollignon} from "./polyhedral/collignon.js";
-export {default as geoPolyhedralWaterman} from "./polyhedral/waterman.js";
+export {default as geoPolyhedralButterfly} from "./butterfly.js";
+export {default as geoPolyhedralCollignon} from "./collignon.js";
+export {default as geoPolyhedralWaterman} from "./waterman.js";
 export {default as geoPolyhedralVoronoi} from "./polyhedral/voronoi.js";
-export {default as geoDodecahedral} from "./polyhedral/dodecahedral.js";
+export {default as geoDodecahedral} from "./dodecahedral.js";
 export {default as geoCox, coxRaw as geoCoxRaw} from "./cox.js";
 export {default as geoTetrahedralLee, leeRaw as geoLeeRaw} from "./tetrahedralLee.js";
 export {default as geoGrayFullerRaw} from "./grayfuller.js";
